Enable smooth scrolling
Added support for animated scrolling. Animation occurs both when
stepAhead() / stepBack() functions are called, but also when an anchor
link is clicked (uses new scrollToAnchor function). In both cases the
animation is controlled by a new helper function, smoothStep, which
accepts a single parameter of new scroll value (an integer).

This solution uses a mixture of of jQuery and plain old JS to achieve
the desired behavior.